requirements.txt
APScheduler==3.10.1
beautifulsoup4~=4.11.1
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
fasttext-wheel==0.9.2
filelock==3.12.2
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
Jinja2==3.1.2
lit==16.0.6
loguru==0.7.0
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==9.3.0
protobuf==3.20.0
psycopg2-binary==2.9.6
pybind11==2.10.4
pydantic~=1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz~=2022.7
PyYAML==6.0
regex==2023.6.3
requests~=2.28.1
safetensors==0.3.1
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.4.1
SQLAlchemy~=1.4.46
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers~=4.29.2
triton==2.0.0
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3

fasttext~=0.9.2
emoji~=2.7.0